+++ kwplot/auto_backends.py
@@ -108,6 +108,17 @@ def autompl(verbose=0, recheck=False, force=None):
             time).
         force (str, default=None): backend to force to or "auto"
 
+    Checks:
+        export QT_DEBUG_PLUGINS=1
+        xdoctest -m kwplot.auto_backends autompl --check
+        KWPLOT_UNSAFE=1 xdoctest -m kwplot.auto_backends autompl --check
+        KWPLOT_UNSAFE=0 xdoctest -m kwplot.auto_backends autompl --check
+
+    Example:
+        >>> # xdoctest +REQUIRES(--check)
+        >>> plt = autoplt(verbose=1)
+        >>> plt.figure()
+
     References:
         https://stackoverflow.com/questions/637005/check-if-x-server-is-running
     """
@@ -157,6 +168,28 @@ def autompl(verbose=0, recheck=False, force=None):
                     This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.
 
                     Available platform plugins are: eglfs, linuxfb, minimal, minimalegl, offscreen, vnc, wayland-egl, wayland, wayland-xcomposite-egl, wayland-xcomposite-glx, webgl, xcb.
+
+
+                UPDATE 2021-01-04:
+
+                    By setting
+
+                    export QT_DEBUG_PLUGINS=1
+
+                    I was able to look at more debug information. It turns out
+                    that it was grabbing the xcb plugin from the opencv-python
+                    package. I uninstalled that package and then installed
+                    opencv-python-headless which does not include an xcb
+                    binary. However, now the it is missing "libxcb-xinerama".
+
+                    May be able to do something with:
+                        conda install -c conda-forge xorg-libxinerama
+
+                        # But that didnt work I had to
+                        pip uninstall PyQt5
+
+                        # This seems to work correctly
+                        conda install -c anaconda pyqt
                 """
                 if ub.modname_to_modpath('PyQt5'):
                     try:
@@ -166,6 +199,41 @@ def autompl(verbose=0, recheck=False, force=None):
                         backend = 'agg'
                     else:
                         backend = 'Qt5Agg'
+
+                        KWPLOT_UNSAFE = os.environ.get('KWPLOT_UNSAFE', '')
+                        TRY_AVOID_CRASH = KWPLOT_UNSAFE.lower() not in ['1', 'true', 'yes']
+
+                        if TRY_AVOID_CRASH and ub.LINUX:
+                            # HOLD UP. Lets try to avoid a crash.
+                            if 'cv2' in sys.modules:
+                                from os.path import dirname, join, exists
+                                cv2 = sys.modules['cv2']
+                                cv2_mod_dpath = dirname(cv2.__file__)
+                                cv2_lib_dpath = join(cv2_mod_dpath, 'qt/plugins/platforms')
+                                cv2_qxcb_fpath = join(cv2_lib_dpath, 'libqxcb.so')
+
+                                qt_mod_dpath = dirname(QtCore.__file__)
+                                qt_lib_dpath = join(qt_mod_dpath, 'Qt/plugins/platforms')
+                                qt_qxcb_fpath = join(qt_lib_dpath, 'libqxcb.so')
+
+                                if exists(cv2_qxcb_fpath) and exists(qt_qxcb_fpath):
+                                    # Can we use ldd to make the test better?
+                                    import warnings
+                                    warnings.warn(ub.paragraph(
+                                        '''
+                                        Autompl has detected libqxcb in PyQt
+                                        and cv2.  Falling back to agg to avoid
+                                        a potential crash. This can be worked
+                                        around by installing
+                                        opencv-python-headless instead of
+                                        opencv-python.
+
+                                        Disable this check by setting the
+                                        environ KWPLOT_UNSAFE=1
+                                        '''
+                                    ))
+                                    backend = 'agg'
+
                 elif ub.modname_to_modpath('PyQt4'):
                     try:
                         import Qt4Agg  # NOQA
@@ -178,6 +246,15 @@ def autompl(verbose=0, recheck=False, force=None):
                     backend = 'agg'
 
             set_mpl_backend(backend, verbose=verbose)
+
+        if 0:
+            # TODO:
+            # IF IN A NOTEBOOK, BE SURE TO SET INLINE BEHAVIOR
+            # THIS EFFECTIVELY REPRODUCES THE %matplotlib inline behavior
+            # BUT USING AN ACTUAL PYTHON FUNCTION
+            shell = _current_ipython_session()
+            if shell:
+                shell.enable_matplotlib('inline')
 
         _AUTOMPL_WAS_RUN = True
 
@@ -197,15 +274,19 @@ class BackendContext(object):
     Context manager that ensures a specific backend, but then reverts after the
     context has ended.
 
-    Ignore:
+    Checks:
+        xdoctest -m kwplot.auto_backends BackendContext --check
+
+    Example:
+        >>> # xdoctest +REQUIRES(--check)
         >>> from kwplot.auto_backends import *  # NOQA
         >>> import matplotlib as mpl
         >>> import kwplot
         >>> print(mpl.get_backend())
-        >>> kwplot.autompl(force='auto')
-        >>> print(mpl.get_backend())
-        >>> fig1 = kwplot.figure(fnum=3)
-        >>> print(mpl.get_backend())
+        >>> #kwplot.autompl(force='auto')
+        >>> #print(mpl.get_backend())
+        >>> #fig1 = kwplot.figure(fnum=3)
+        >>> #print(mpl.get_backend())
         >>> with BackendContext('agg'):
         >>>     print(mpl.get_backend())
         >>>     fig2 = kwplot.figure(fnum=4)
@@ -215,12 +296,51 @@ class BackendContext(object):
     def __init__(self, backend):
         self.backend = backend
         self.prev = None
+        self._prev_backend_was_loaded = 'matplotlib.pyplot' in sys.modules
 
     def __enter__(self):
         import matplotlib as mpl
         self.prev = mpl.get_backend()
+
+        if self.prev == 'Qt5Agg':
+            # Hack for the case where our default matplotlib backend is Qt5Agg
+            # but we don't have Qt bindings available. (I think this may be a
+            # configuration error on my part). Either way, its easy to test for
+            # and fix. If the default backend is Qt5Agg, but importing the
+            # bindings causes an error, just set the default to agg, which will
+            # supress the warnings.
+            try:
+                from matplotlib.backends.qt_compat import QtGui  # NOQA
+            except ImportError:
+                # TODO: should we try this instead?
+                # mpl.rcParams['backend_fallback']
+                self.prev = 'agg'
+
         set_mpl_backend(self.backend)
 
     def __exit__(self, *args):
         if self.prev is not None:
-            set_mpl_backend(self.prev)
+            """
+            Note: 2021-01-07
+                Running on in an ssh-session (where in this case ssh did
+                not have an X server, but an X server was running
+                elsewhere) we got this error.
+
+                ImportError: Cannot load backend 'Qt5Agg' which requires the
+                'qt5' interactive framework, as 'headless' is currently running
+
+                when using BackendContext('agg')
+
+                This is likely because the default was Qt5Agg, but it was not
+                loaded. We switched to agg just fine, but when we switched back
+                it tried to load Qt5Agg, which was not available and thus it
+                failed.
+            """
+            try:
+                # Note
+                set_mpl_backend(self.prev)
+            except Exception:
+                if self._prev_backend_was_loaded:
+                    # Only propogate the error if we had explicitly used pyplot
+                    # beforehand. Note sure if this is the right thing to do.
+                    raise